+++ resources/views/product/index.php
@@ -10,8 +10,10 @@
             <td><?= $i ?></td>
             <td><?= $product['name'] ?></td>
             <td><img src="<?= BASEURL ?>/assets/img/uploads/<?= $product['photo'] ?>" alt="<?= $product['photo'] ?>" width="300"></td>
-            <td><a href="<?= BASEURL ?>/product/edit/<?= $product['id'] ?>">EDIT</a></td>
-            <td><a href="<?= BASEURL ?>/product/delete/<?= $product['id'] ?>" onclick="return confirm('Yakin ingin menghapus data ini?')">DELETE</a></td>
+            <?php if (auth()->role === 'admin') : ?>
+                <td><a href="<?= BASEURL ?>/product/edit/<?= $product['id'] ?>">EDIT</a></td>
+                <td><a href="<?= BASEURL ?>/product/delete/<?= $product['id'] ?>" onclick="return confirm('Yakin ingin menghapus data ini?')">DELETE</a></td>
+            <?php endif ?>
         </tr>
         <?php $i++ ?>
     <?php endforeach ?>
